Created bank_users.borrower_dashboard.borrower_loan_close.view_profile; Edited bank_users.borrower_dashboard.borrower_loan_close.RowTemplate15, bank_users.borrower_dashboard.application_tracker.view_profile and bank_users.borrower_dashboard.application_tracker.RowTemplate1; Deleted bank_users.borrower_dashboard.borrower_loan_close.view_profile_copy_copy
