+++ .anvil_editor.yaml
@@ -108,7 +108,6 @@ unique_ids:
     lendor_registration_form.dashboard.vsn: '1699967294797596075787674.3921'
     lendor_registration_form.dashboard.vlo: '1699961316448706281090869.9006'
     borrower_registration_form.star_1_borrower_registration_form_begin_3.star_1_borrower_registration_form_begin_3b_business_3: '1700126739765126145778650.8795'
-    bank_users.borrower_dashboard.borrower_loan_close.view_profile_copy_copy: '1701149621405989016570674.1494'
     lendor_registration_form.dashboard.ld: '1699960450618538570671683.908'
     lendor_registration_form.dashboard.vblr.Borr_loan_request.RowTemplate11: '1700543062684858008325535.9921'
     admin.dashboard.manage_products.manage_producs1: '1700713086879771707093045.5188'
@@ -148,6 +147,7 @@ unique_ids:
     borrower_registration_form.star_1_borrower_registration_form_begin_3.star_1_borrower_registration_form_begin_3f: '170003767984838881190201.539795'
     admin.dashboard.lenders.RowTemplate4: '1700737487916312936831844.4892'
     admin.dashboard.admin_teams.view_profile: '1699528208717177568108194.0108'
+    bank_users.borrower_dashboard.borrower_loan_close.view_profile: '1701149621405989016570674.1494'
     admin.dashboard.manage_products.add_product_categories_and_groups: '1701090290692291670903022.0623'
     lendor_registration_form.dashboard.opbal: '1699956535223911918554378.0013'
     admin.dashboard.loan_management.closed_loans.view_profile_5: '1701079991010941258272977.0447'
