+++ client_code/bank_users/borrower_dashboard/application_tracker/RowTemplate1/__init__.py
@@ -18,5 +18,5 @@ class RowTemplate1(RowTemplate1Template):
 
   def loan_status_click(self, **event_args):
     """This method is called when the link is clicked"""
-    value=self.loan_status
-    open_form('bank_users.borrower_dashboard.application_tracker.view_profile',value=value)
+    selected_row=self.item
+    open_form('bank_users.borrower_dashboard.application_tracker.view_profile',selected_row=selected_row)
