+++ client_code/bank_users/borrower_dashboard/borrower_loan_close/view_profile/__init__.py
@@ -1,0 +1,26 @@
+from ._anvil_designer import view_profileTemplate
+from anvil import *
+import anvil.server
+import anvil.google.auth, anvil.google.drive
+from anvil.google.drive import app_files
+import anvil.users
+import anvil.tables as tables
+import anvil.tables.query as q
+from anvil.tables import app_tables
+from .. import borrower_main_form_module as main_form_module
+
+class view_profile(view_profileTemplate):
+  def __init__(self, **properties):
+    self.user_id=main_form_module.userId
+    #self.user_id=1000
+    self.init_components(**properties)
+
+    # Any code you write here will run before the form opens.
+   
+
+  def button_1_copy_click(self, **event_args):
+    open_form('bank_users.borrower_dashboard')
+
+  def button_1_click(self, **event_args):
+    """This method is called when the button is clicked"""
+    open_form('bank_users.borrower_dashboard.borrower_loan_close')
