+++ client_code/bank_users/borrower_dashboard/borrower_loan_close/view_profile/form_template.yaml
@@ -30,7 +30,7 @@ components:
     grid_position: PNDNDM,TKYEBB
 - name: label_1_copy_1_copy_1
   properties: {role: null, align: center, tooltip: '', border: '', foreground: '',
-    visible: true, text: View Profile, font_size: null, font: '', spacing_above: small,
+    visible: true, text: Closed Profile, font_size: null, font: '', spacing_above: small,
     icon_align: left, spacing_below: small, italic: false, background: '', bold: true,
     underline: false, icon: ''}
   type: Label
